+++ agent/src/test/java/com/navercorp/pinpoint/plugin/mybatis/SqlSessionTemplateITBase.java
@@ -1,0 +1,7 @@
+package com.navercorp.pinpoint.plugin.mybatis;
+
+/**
+ * @author HyunGil Jeong
+ */
+public class SqlSessionTemplateITBase {
+}
